requirements.txt
anthropic>=0.25.0
python-dotenv>=1.0.0
notion-client>=2.0.0
gspread>=6.0.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
python-telegram-bot>=20.0
httpx>=0.27.0
google-genai>=0.3.0
youtube-transcript-api>=0.6.0
